Add files via upload
It was prepared a separate file to perform a sentiment analysis based on "construction subject". 

It was basically to load the credentials using data from Reddit as it was learnt in programming class.  It was able to scrape data from that topic (construction) and then created a data frame. It was also performed some lambda functions to extract data from the text